% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{colNormalize}
\alias{colNormalize}
\title{Normalize each column of the input matrix by the column sum}
\usage{
colNormalize(x, scaleFactor = NULL, log = FALSE)
}
\arguments{
\item{x}{Feature by observation matrix.}

\item{scaleFactor}{Multiplier on normalized data. Default \code{NULL}.}

\item{log}{Logical. Whether to take log1p transformation after scaling.
Default \code{FALSE}}
}
\value{
Normalized matrix of the same size
}
\description{
Normalize each column of the input matrix by the column sum
}
\examples{
rnaNorm <- colNormalize(rnaRaw)
}
