% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{plotBinary}
\alias{plotBinary}
\alias{plotBinary.simMat}
\alias{plotBinary.default}
\alias{plotBinary.liger}
\title{Create binary plots}
\usage{
plotBinary(object, ...)

\method{plotBinary}{simMat}(
  object,
  axisPortion = TRUE,
  dotSize = 0.6,
  dotColor = "grey60",
  densLinewidth = 0.8,
  labelColors = c("#3B4992FF", "#EE0000FF"),
  distMethod = attributes(object)$method,
  title = NULL,
  ...
)

\method{plotBinary}{default}(
  object,
  clusterVar,
  vertices,
  method = c("euclidean", "cosine", "pearson", "spearman"),
  force = FALSE,
  sigma = 4,
  scale = TRUE,
  splitCluster = FALSE,
  clusterTitle = TRUE,
  dotColor = "grey60",
  ...
)

\method{plotBinary}{liger}(object, clusterVar, features = NULL, useDatasets = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{axisPortion}{Whether to show axis value by proportion, so that the
coordinates of each dot sum up to 100. Default \code{TRUE}.}

\item{dotSize, dotColor}{Dot aesthetics passed to
\code{\link[ggplot2]{geom_point}}. Default \code{0.6} and \code{"grey60"}.}

\item{densLinewidth}{Density plot line aesthetic. Default \code{0.8}.}

\item{labelColors}{Color of the axis lines and vertex labels. Default
\code{c("#3B4992FF", "#EE0000FF")} (blue and red).}

\item{distMethod}{Method name used for calculating the dist.matrix object.}

\item{title}{Title text of the plot. Default \code{NULL}.}

\item{clusterVar}{A vector/factor assigning the cluster variable to each
column of the matrix object. Or a character scalar selecting a cell metadata
variable from container object.}

\item{vertices}{A vector of TWO values specifying the clusters as the
terminals.}

\item{method}{Distance calculation method. Default \code{"euclidean"}.
Choose from \code{"euclidean"}, \code{"cosine"}, \code{"pearson"},
\code{"spearman"}.}

\item{force}{Whether to force calculate the distance when more then 500
features are detected, which is generally not recommended. Default
\code{FALSE}.}

\item{sigma}{Gaussian kernel parameter that controls the effect of variance.
Only effective when using a distance metric (i.e. \code{method} is
\code{"euclidian"} or \code{"cosine"}). Larger value tighten the dot
spreading on figure. Default \code{4}.}

\item{scale}{Whether to min-max scale the distance matrix by clusters.
Default \code{TRUE}.}

\item{splitCluster}{Logical, whether to return a list of plots where each
contains cells belonging to only one cluster. Default \code{FALSE}.}

\item{clusterTitle}{If \code{splitCluster = TRUE}, whether to title each
subplot by the name of each cluster. Default \code{TRUE}.}

\item{features}{For container object methods. Valid row subsetting index that
selects features. Default \code{NULL}.}

\item{useDatasets}{For liger method, select datasets where the distance
calculation should only be limited within this range. Default \code{NULL}
uses all datasets.}
}
\value{
For distMatrix method, a ggplot object. For other methods, a ggplot
object when \code{splitCluster = FALSE}, or a list of ggplot objects when
\code{splitCluster = TRUE}.
}
\description{
Create binary plots that show similarity between single cells and
selected terminals in a 2D space. The two vertices are placed at the left and
right of a 2D plot where x-axis shows the similarity. Y-axis is jittered for
clear view. A density curve is added for indicating the distribution.
}
\examples{
rnaLog <- colNormalize(rnaRaw, 1e4, TRUE)
gene <- selectTopFeatures(rnaRaw, rnaCluster, c("RE", "OS"))
plotBinary(rnaLog[gene, ], rnaCluster, c("RE", "OS"))
}
