% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternary.R
\name{plotQuaternary}
\alias{plotQuaternary}
\alias{plotQuaternary.simMat}
\alias{plotQuaternary.default}
\alias{plotQuaternary.Seurat}
\title{Create quaternary simplex plots}
\usage{
plotQuaternary(object, ...)

\method{plotQuaternary}{simMat}(
  object,
  veloMat = NULL,
  nGrid = 10,
  radius = 0.2,
  dotSize = 0.6,
  dotColor = "grey60",
  labelColors = c("#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF"),
  arrowLinewidth = 0.6,
  title = NULL,
  theta = 0,
  phi = 0,
  ...
)

\method{plotQuaternary}{default}(
  object,
  clusterVar,
  vertices,
  veloGraph = NULL,
  method = c("euclidean", "cosine", "pearson", "spearman"),
  force = FALSE,
  sigma = 0.08,
  scale = TRUE,
  splitCluster = FALSE,
  clusterTitle = TRUE,
  dotColor = "grey60",
  ...
)

\method{plotQuaternary}{Seurat}(object, features = NULL, slot = "data", assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{veloMat}{Aggregated velocity matrix. Output of \code{aggrVeloGraph}.}

\item{nGrid}{Number of grids along the x-axis of the tetrahedron
triangle. Default \code{10}.}

\item{radius}{Arrow length of unit velocity. Lower this when arrows point
outside of the tetrahedron. Default \code{0.2}.}

\item{dotSize, dotColor}{Dot aesthetics. Default \code{0.6} and
\code{"grey60"}.}

\item{labelColors}{Colors of the vertex labels. Default
\code{c("#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF")} (blue, red,
green and purple).}

\item{arrowLinewidth}{Arrow aesthetics. Default \code{0.6}.}

\item{title}{Title text of the plot. Default \code{NULL}.}

\item{theta, phi}{Numeric scalar. The angles defining the viewing direction.
\code{theta} gives the azimuthal direction and \code{phi} the colatitude.
Default \code{-40} and \code{-10}.}

\item{clusterVar}{A vector/factor assigning the cluster variable to each
column of the matrix object. Or a character scalar selecting a cell metadata
variable from container object.}

\item{vertices}{A vector of TWO values specifying the clusters as the
terminals.}

\item{veloGraph}{Cell x cell dgCMatrix object containing velocity
information. Shows velocity grid-arrow layer when specified. Default
\code{NULL} does not show velocity.}

\item{method}{Distance calculation method. Default \code{"euclidean"}.
Choose from \code{"euclidean"}, \code{"cosine"}, \code{"pearson"},
\code{"spearman"}.}

\item{force}{Whether to force calculate the distance when more then 500
features are detected, which is generally not recommended. Default
\code{FALSE}.}

\item{sigma}{Gaussian kernel parameter that controls the effect of variance.
Only effective when using a distance metric (i.e. \code{method} is
\code{"euclidian"} or \code{"cosine"}). Larger value tighten the dot
spreading on figure. Default \code{0.08}.}

\item{scale}{Whether to min-max scale the distance matrix by clusters.
Default \code{TRUE}.}

\item{splitCluster}{Logical, whether to return a list of plots where each
contains cells belonging to only one cluster. Default \code{FALSE}.}

\item{clusterTitle}{If \code{splitCluster = TRUE}, whether to title each
subplot by the name of each cluster. Default \code{TRUE}.}

\item{features}{For container object methods. Valid row subsetting index that
selects features. Default \code{NULL} uses all available features.}

\item{slot}{For Seurat method, choose from \code{"data"},
\code{"scale.data"} or \code{"counts"}. Default \code{"data"}.}

\item{assay}{For Seurat method, the specific assay to get data from. Default
\code{NULL} to the default assay.}
}
\value{
For "simMat" method, a "plist" (plot3D package product) object. For
other methods, a "plist" object when \code{splitCluster = FALSE}, or a list
of "plist" objects when \code{splitCluster = TRUE}. A "plist" object can be
viewed with \code{print()}, \code{show()} or a direct run of the object
variable name in interactive console.
}
\description{
Create quaternary plots that show similarity between single cells and
selected four terminals in a tetrahedron space.

A dynamic rotating view in a GIF image file can be created with
\code{\link{writeQuaternaryGIF}}. Package \code{magick} must be installed in
advance. Linux users may refer to this
\href{https://cran.r-project.org/web/packages/magick/vignettes/intro.html#Build_from_source}{installation guide}.
}
\examples{
rnaNorm <- colNormalize(rnaRaw)
gene <- selectTopFeatures(rnaNorm, rnaCluster, c("RE", "OS", "CH", "ORT"))
rnaLog <- colNormalize(rnaRaw, 1e4, TRUE)
plotQuaternary(rnaLog[gene, ], rnaCluster, c("RE", "OS", "CH", "ORT"))

# Seurat example
if (FALSE) {
    library(Seurat)
    srt <- CreateSeuratObject(rnaRaw)
    Idents(srt) <- rnaCluster
    geneSel <- selectTopFeatures(rnaRaw, rnaCluster,
                                 vertices = c("OS", "RE", "CH", "ORT"))
    srt <- NormalizeData(srt)
    plotQuaternary(srt, features = geneSel, vertices = c("OS", "RE", "CH", "ORT"))
}
}
