% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{plotTernary}
\alias{plotTernary}
\alias{plotTernary.distMatrix}
\alias{plotTernary.default}
\alias{plotTernary.liger}
\title{Create ternary plots}
\usage{
plotTernary(object, ...)

\method{plotTernary}{distMatrix}(
  object,
  veloMat = NULL,
  axisPortion = TRUE,
  dotSize = 0.6,
  dotColor = "grey60",
  labelColors = c("#EE0000FF", "#3B4992FF", "#008B45FF"),
  distMethod = attributes(object)$method,
  title = NULL,
  nGrid = 10,
  radius = 0.1,
  arrowLinewidth = 0.25,
  ...
)

\method{plotTernary}{default}(
  object,
  clusterVar,
  vertices,
  method = c("euclidean", "cosine", "pearson", "spearman"),
  veloGraph = NULL,
  normCluster = FALSE,
  scale = TRUE,
  splitCluster = FALSE,
  clusterTitle = TRUE,
  ...
)

\method{plotTernary}{liger}(object, clusterVar, features = NULL, useDatasets = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{veloMat}{Aggregated velocity matrix. Output of \code{aggrVeloGraph}.}

\item{axisPortion}{Whether to show axis value by proportion, so that the
coordinates of each dot sum up to 100. Default \code{TRUE}.}

\item{dotSize, dotColor}{Dot aesthetics passed to
\code{\link[ggplot2]{geom_point}}. Default \code{0.6} and \code{"grey60"}.}

\item{labelColors}{Colors of the axis lines and vertex labels.
Default \code{c("#EE0000FF", "#3B4992FF", "#008B45FF")} (red, blue and green)}

\item{distMethod}{Method name used for calculating the dist.matrix object.}

\item{title}{Title text of the plot. Default \code{NULL}.}

\item{nGrid}{Number of grids along the bottom side of the equilateral
triangle. Default \code{10}.}

\item{radius}{Arrow length of unit velocity. Lower this when arrows point
outside of the coordinate. Default \code{0.1}.}

\item{arrowLinewidth}{Arrow aesthetics. Default \code{0.25}.}

\item{clusterVar}{A vector/factor assigning the cluster variable to each
column of the matrix object. Or a character scalar selecting a cell metadata
variable from container object.}

\item{vertices}{A vector of THREE values specifying the clusters as the
terminals.}

\item{method}{Distance calculation method. Default \code{"euclidean"}.
Choose from \code{"euclidean"}, \code{"cosine"}, \code{"pearson"},
\code{"spearman"}.}

\item{veloGraph}{Cell x cell dgCMatrix object containing velocity
information. Shows velocity grid-arrow layer when specified. Default
\code{NULL} does not show velocity.}

\item{normCluster}{Whether to normalize the distance matrix by clusters.
See Details. Default \code{FALSE}.}

\item{scale}{Whether to min-max scale the distance matrix by clusters.
Default \code{TRUE}.}

\item{splitCluster}{Logical, whether to return a list of plots where each
contains cells belonging to only one cluster. Default \code{FALSE}.}

\item{clusterTitle}{If \code{splitCluster = TRUE}, whether to title each
subplot by the name of each cluster. Default \code{TRUE}.}

\item{features}{For container object methods. Valid row subsetting index that
selects features. Default \code{NULL}.}

\item{useDatasets}{For liger method, select datasets where the distance
calculation should only be limited within this range. Default \code{NULL}
uses all datasets.}
}
\value{
For distMatrix method, a ggplot object. For other methods, a ggplot
object when \code{splitCluster = FALSE}, or a list of ggplot objects when
\code{splitCluster = TRUE}.
}
\description{
Create ternary plots that show similarity between single cells and
selected terminals in a triangle 2D space.
}
\details{
\bold{Argument inheritance} - For matrix/dgCMatrix input object, we first
calculate the similarity matrix and obtain a \code{dist.matrix} object. Then
we call the \code{plotTernary.dist.matrix} method. For data container objects
(e.g. liger), we obtain the correct data matrix first and then call the
matrix/dgCMatrix method. Therefore, the arguments inherits as the flow
described above.

\bold{The calculation of similarity matrix} - TODO
}
