% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternary.R
\name{print.quatPlot}
\alias{print.quatPlot}
\title{Show plist object produced with plot3D package}
\usage{
\method{print}{quatPlot}(x, ...)
}
\arguments{
\item{x}{quatPlot object, returned by \code{\link{plotQuaternary}} when
\code{interactive = FALSE}.}

\item{...}{Graphic parameters passed to \code{\link{plot}}. \code{mar} is
pre-specified.}
}
\value{
No return value. It displays the plot described in a 'plist' object
returned by \code{\link{plotQuaternary}}, internally created by package
'plot3D'.
}
\description{
Show plist object produced with plot3D package
}
\examples{
gene <- selectTopFeatures(
    x = rnaRaw,
    clusterVar = rnaCluster,
    vertices = c("RE", "OS", "CH", "ORT")
)
quat <- plotQuaternary(
    x = rnaRaw,
    clusterVar = rnaCluster,
    vertices = c("RE", "OS", "CH", "ORT"),
    features = gene,
    interactive = FALSE
)
quat; print(quat)
}
