% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaCluster}
\alias{rnaCluster}
\title{Major cell type annotation of the example mouse bone marrow data}
\format{
\code{factor} object
}
\source{
https://www.nature.com/articles/s41467-023-38034-2
}
\usage{
rnaCluster
}
\description{
Major cell type annotation of the example mouse bone marrow data
}
\references{
Matsushita, Y., Liu, J., Chu, A.K.Y. et al. Bone marrow endosteal
stem cells dictate active osteogenesis and aggressive tumorigenesis. Nat
Commun 14, 2383 (2023).
}
\keyword{datasets}
