% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaVelo}
\alias{rnaVelo}
\title{Velocity graph of the example mouse bone marrow data}
\format{
\link[Matrix]{dgCMatrix-class} object
}
\source{
https://www.nature.com/articles/s41467-023-38034-2
}
\usage{
rnaVelo
}
\description{
Velocity graph of the example mouse bone marrow data
}
\references{
Matsushita, Y., Liu, J., Chu, A.K.Y. et al. Bone marrow endosteal
stem cells dictate active osteogenesis and aggressive tumorigenesis. Nat
Commun 14, 2383 (2023).
}
\keyword{datasets}
