% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTopFeatures.R
\name{selectTopFeatures}
\alias{selectTopFeatures}
\alias{selectTopFeatures.default}
\alias{selectTopFeatures.Seurat}
\alias{selectTopFeatures.SingleCellExperiment}
\title{Pick top differentially presented features for similarity calculation}
\usage{
selectTopFeatures(x, clusterVar, vertices, ...)

\method{selectTopFeatures}{default}(
  x,
  clusterVar,
  vertices,
  nTop = 30,
  lfcThresh = 0.1,
  returnStats = FALSE,
  ...
)

\method{selectTopFeatures}{Seurat}(
  x,
  clusterVar = NULL,
  vertices,
  assay = NULL,
  slot = "data",
  ...
)

\method{selectTopFeatures}{SingleCellExperiment}(
  x,
  clusterVar = NULL,
  vertices,
  assay.type = "normcounts",
  ...
)
}
\arguments{
\item{x}{Dense or sparse matrix, observation per column. Must be library size
normalized but not log-transformed. Alternatively, a \code{Seurat} object or
a \code{SingleCellExperiment} object with library size normalized data
available.}

\item{clusterVar}{A vector/factor assigning the cluster variable to each
column of the matrix object. For "Seurat" method, leave \code{NULL} for using
default "Idents", or can also be a variable in \code{meta.data} slot. For
"SingleCellExperiment" method, leave \code{NULL} for using "colLabels", or
can be a variable in \code{colData} slot.}

\item{vertices}{Vector of cluster names that will be used for plotting. Or a
named list that groups clusters as a terminal vertex. There must not be any
overlap between groups.}

\item{...}{Arguments passed to methods.}

\item{nTop}{Number of top differentially presented features per cluster.
Default \code{30}.}

\item{lfcThresh}{Threshold on log fold-change to identify up-regulated
features. Default \code{0.1}.}

\item{returnStats}{Logical. Whether to return the full statistics table
rather then returning the selected genes. Default \code{FALSE}}

\item{assay}{Assay name of the Seurat object to be used. Default \code{NULL}.}

\item{slot}{Slot name of the Seurat object to be used. Default \code{"data"}.}

\item{assay.type}{Assay name of the SingleCellExperiment object to be used.}
}
\value{
When \code{returnStats = FALSE} (default), a character vector of at
most \code{length(unique(vertices))*nTop} feature names. When
\code{returnStats = TRUE}, a data.frame of wilcoxon rank sum test statistics.
}
\description{
Performs wilcoxon rank sum test on input matrix. Comparisons are formed by
observations in each group, as defined by \code{clusterVar}, against the
other observations. The U-Statistics (\code{statistic}), p-value
(\code{pval}) and adjusted p-value (\code{padj}), together with average
presence in group (\code{avgExpr}), log fold-change (\code{logFC}), AUC
(\code{auc}), percentage in group (\code{pct_in}) and percentage out of group
(\code{pct_out}) will be calculated.

Top features are selected by sorting primarily on adjusted p-value, and
secondarily on log fold-change, after filtering for up-regulated features.
}
\examples{
rnaNorm <- colNormalize(rnaRaw)
selectTopFeatures(rnaNorm, rnaCluster, c("OS", "RE"))

# Seurat example
if (FALSE) {
    library(Seurat)
    srt <- CreateSeuratObject(rnaRaw)
    Idents(srt) <- rnaCluster
    srt <- colNormalize(srt)
    gene <- selectTopFeatures(srt, vertices = c("OS", "RE"))
}

# SingleCellExperiment example
if (FALSE) {
    library(SingleCellExperiment)
    sce <- SingleCellExperiment(assays = list(counts = rnaRaw))
    colLabels(sce) <- rnaCluster
    sce <- colNormalize(sce)
    gene <- selectTopFeatures(sce, vertices = c("OS", "RE"))
}
}
