% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternary.R
\name{show,plist-method}
\alias{show,plist-method}
\alias{print.plist}
\title{Show plist object produced with plot3D package}
\usage{
\S4method{show}{plist}(object)

\method{print}{plist}(x, ...)
}
\arguments{
\item{object, x}{plist object}

\item{...}{Graphic parameters passed to \code{\link{plot}}. \code{mar} is
pre-specified.}
}
\value{
No return value. It displays the plot described in a 'plist' object
returned by \code{\link{plotQuaternary}}, internally created by package
'plot3D'.
}
\description{
Show plist object produced with plot3D package
}
\examples{
gene <- selectTopFeatures(rnaRaw, rnaCluster, c("RE", "OS", "CH", "ORT"))
plistObj <- plotQuaternary(rnaRaw, rnaCluster, c("RE", "OS", "CH", "ORT"), gene)
print(plistObj)
# equivalent to
show(plistObj)
}
