% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternary.R
\name{show,plist-method}
\alias{show,plist-method}
\alias{print.plist}
\title{Show plist object produced with plot3D package}
\usage{
\S4method{show}{plist}(object)

\method{print}{plist}(x, ...)
}
\arguments{
\item{object, x}{plist object}

\item{...}{Not used.}
}
\description{
Show plist object produced with plot3D package
}
