% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{plotTernary}
\alias{plotTernary}
\alias{plotTernary.simMat}
\alias{plotTernary.default}
\title{Create ternary plots}
\usage{
plotTernary(object, ...)

\method{plotTernary}{simMat}(
  object,
  title = NULL,
  veloMat = NULL,
  nGrid = 10,
  radius = 0.1,
  dotSize = 0.6,
  dotColor = "grey60",
  labelColors = c("#3B4992FF", "#EE0000FF", "#008B45FF"),
  vertexLabelSize = 5,
  vertexLabelDrift = 0.03,
  axisBreak = 5,
  axisTextSize = 4,
  axisTextDrift = 0.02,
  gridLineAlpha = 0.6,
  arrowLinewidth = 0.25,
  titleSize = 14,
  equilateral = TRUE,
  margin = 0.1,
  ...
)

\method{plotTernary}{default}(
  object,
  clusterVar,
  vertices,
  method = c("euclidean", "cosine", "pearson", "spearman"),
  veloGraph = NULL,
  force = FALSE,
  sigma = 4,
  scale = TRUE,
  splitCluster = FALSE,
  clusterTitle = TRUE,
  dotColor = "grey60",
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{title}{Title text of the plot. Default \code{NULL}.}

\item{veloMat}{Aggregated velocity matrix. Output of \code{aggrVeloGraph}.}

\item{nGrid}{Number of grids along the bottom side of the equilateral
triangle. Default \code{10}.}

\item{radius}{Arrow length of unit velocity. Lower this when arrows point
outside of the coordinate. Default \code{0.1}.}

\item{dotSize, dotColor}{Dot aesthetics passed to
\code{\link[ggplot2]{geom_point}}. Default \code{0.6} and \code{"grey60"}.}

\item{labelColors}{Colors of the axis lines and vertex labels.
Default \code{c("#3B4992FF", "#EE0000FF", "#008B45FF")} (blue, red and green)}

\item{vertexLabelSize, vertexLabelDrift}{Adjustment on the three vertex text
labels. Drift means the distance that the labels should be moved against the
center of the plot. Default size \code{5}, drifted distance \code{0.03}.}

\item{axisBreak}{Number of breaks to be labeled along axis. Default
\code{5}.}

\item{axisTextSize, axisTextDrift}{Similar to the vertex adjustment applied
to the text label along the axis breaks. Default size \code{4}, drifted
distance \code{0.02}.}

\item{gridLineAlpha}{Transparency of background grid lines. Default
\code{0.6}.}

\item{arrowLinewidth}{Arrow aesthetics. Default \code{0.25}.}

\item{titleSize}{Size of title text. Default \code{14}.}

\item{equilateral}{Logical, whether to always display the triangle as
equilateral. Default \code{TRUE}.}

\item{margin}{Margin allowed around of the triangle plotting region when
\code{equilateral = TRUE}}

\item{clusterVar}{A vector/factor assigning the cluster variable to each
column of the matrix object. Or a character scalar selecting a cell metadata
variable from container object.}

\item{vertices}{A vector of THREE values specifying the clusters as the
terminals.}

\item{method}{Distance calculation method. Default \code{"euclidean"}.
Choose from \code{"euclidean"}, \code{"cosine"}, \code{"pearson"},
\code{"spearman"}.}

\item{veloGraph}{Cell x cell dgCMatrix object containing velocity
information. Shows velocity grid-arrow layer when specified. Default
\code{NULL} does not show velocity.}

\item{force}{Whether to force calculate the distance when more then 500
features are detected, which is generally not recommended. Default
\code{FALSE}.}

\item{sigma}{Gaussian kernel parameter that controls the effect of variance.
Only effective when using a distance metric (i.e. \code{method} is
\code{"euclidian"} or \code{"cosine"}). Larger value tighten the dot
spreading on figure. Default \code{4}.}

\item{scale}{Whether to min-max scale the distance matrix by clusters.
Default \code{TRUE}.}

\item{splitCluster}{Logical, whether to return a list of plots where each
contains cells belonging to only one cluster. Default \code{FALSE}.}

\item{clusterTitle}{If \code{splitCluster = TRUE}, whether to title each
subplot by the name of each cluster. Default \code{TRUE}.}
}
\value{
For distMatrix method, a ggplot object. For other methods, a ggplot
object when \code{splitCluster = FALSE}, or a list of ggplot objects when
\code{splitCluster = TRUE}.
}
\description{
Create ternary plots that show similarity between single cells and
selected terminals in a triangle 2D space.
}
\details{
\bold{Argument inheritance} - For matrix/dgCMatrix input object, we first
calculate the similarity matrix and obtain a \code{dist.matrix} object. Then
we call the \code{plotTernary.dist.matrix} method. For data container objects
(e.g. liger), we obtain the correct data matrix first and then call the
matrix/dgCMatrix method. Therefore, the arguments inherits as the flow
described above.

\bold{The calculation of similarity matrix} - TODO
}
\examples{
rnaLog <- colNormalize(rnaRaw, 1e4, TRUE)
gene <- selectTopFeatures(rnaRaw, rnaCluster, c("RE", "OS", "CH"))
plotTernary(rnaLog[gene, ], rnaCluster, c("RE", "OS", "CH"))
}
