% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTopFeatures.R
\name{selectTopFeatures}
\alias{selectTopFeatures}
\title{Pick top differentially presented features for similarity calculation}
\usage{
selectTopFeatures(
  x,
  clusterVar,
  vertices,
  nTop = 30,
  normalize = TRUE,
  padjThresh = 0.01,
  returnStats = FALSE
)
}
\arguments{
\item{x}{Dense or sparse matrix, observation per column.}

\item{clusterVar}{Grouping labels. Length must match with \code{ncol(x)}}

\item{vertices}{Vector of cluster names that will be used for plotting.}

\item{nTop}{Number of top differentially presented features per cluster.
Default \code{30}.}

\item{normalize}{Logical. Whether to normalize matrix by column sum (i.e.
library size). Default \code{TRUE} when \code{x} is raw count matrix. Users
using already log-transformed normalized input should turn this to
\code{FALSE} as normalizing when then change the feature ranking.}

\item{padjThresh}{Threshold on adjusted p-value to identify significant
features. Default \code{0.01}.}

\item{returnStats}{Logical. Whether to return the whole statistics table
rather then returning the selected genes. Default \code{FALSE}}
}
\value{
When \code{returnStats = FALSE} (default), a character vector of
\code{length(unique(vertices))*nTop} feature names. When \code{returnStats =
TRUE}, a data.frame of wilcoxon rank sum test statistics.
}
\description{
Performs wilcoxon rank sum test on input matrix. Comparisons are formed by
observations in each group, as defined by \code{clusterVar}, against the
other observations. The U-Statistics (\code{statistic}), p-value
(\code{pval}) and adjusted p-value (\code{padj}), together with average
presence in group (\code{avgExpr}), log fold-change (\code{logFC}), AUC
(\code{auc}), percentage in group (\code{pct_in}) and percentage out of group
(\code{pct_out}) will be calculated.
}
\examples{
selectTopFeatures(rnaRaw, rnaCluster, "RE")
}
